requirements.txt
hikari[speedups]>=2.0.0.dev120
hikari-lightbulb>=2.3.3
hikari-miru>=3.1.2

python-dotenv>=0.21.0

aiohttp==3.8.3
plotly
isodate==0.6.1
psutil
Pillow
# tabulate
feedparser
kaleido==0.2.1
aiohttp-client-cache==0.8.1
beautifulsoup4
lxml
fuzzywuzzy[speedup]
pandas
fast-colorthief
pre-commit


# Putting this at the end coz W*ndows doesn't need it
uvloop